// src/app/core/models/user.model.ts
export class User {
    id: number | undefined;
    firstName: string | undefined;
    lastName: string | undefined;
    email: string | undefined;
    mobileNumber: string | undefined;
    phoneNumber: string | undefined;
    password: string | undefined;
    newPassword: string | undefined;
    verifyPassword: string | undefined;
    profileImage: string | undefined;
    profileImage_CloudinaryPublicId: string | undefined;
    gender: number | undefined;
    emailVerified: boolean | undefined;
    mobileNumberVerified: boolean | undefined;
    userLoginEnabled: boolean | undefined;
    fullName: string | undefined;
    userRoleCode: string | undefined;
    resetPasswordLink: string | undefined;
}